requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
spotipy>=2.23.0
python-dotenv>=1.0.0
